I was trying to make use of ioredis-mock as the testing backend for bull queing library.
So I needed to add a couple more commands which I think that they should be added to ioredis-mock through separate PRs in a more clean way.

But... as I 'm on a hurry right now... I 'm preparing this single PR with the stuff I needed to do here.